refactor(srv): split reducer.rs into per-domain submodules
Pure code-move per docs/specs/PLAN_SRV_REDUCER_MODULARIZATION_2026-05-07.md.
No behaviour change — 866/866 srv tests pass identically post-split.

Pre-split: agentmux-srv/src/reducer.rs at 4385 lines with 28
handlers + 2900 lines of tests in one file. Outlier vs the host
(agentmux-cef/src/reducer/ already modular) and launcher
(agentmux-launcher/src/reducer/ already modular) reducers.

Post-split shape (mirrors agentmux-cef/src/reducer/):

  agentmux-srv/src/reducer.rs           — public update() dispatch + Ctx + tests (3058 lines)
  agentmux-srv/src/reducer/lifecycle.rs — Register, Goodbye (103)
  agentmux-srv/src/reducer/snapshot.rs  — GetSrvSnapshot (50)
  agentmux-srv/src/reducer/workspace.rs — Create/Delete/Rename/UpdateMeta (149)
  agentmux-srv/src/reducer/tab.rs       — Create/Delete/SetActive/Reorder/Bulk/Move/Rename/Meta (490)
  agentmux-srv/src/reducer/block.rs     — Create/Delete/Move/UpdateMeta (169)
  agentmux-srv/src/reducer/layout.rs    — SetFocused/SetMagnified + 4 LayoutXxx arms (324)
  agentmux-srv/src/reducer/window.rs    — CreateWindow/Close/SwitchWorkspace (103)

Mechanics:
  - Each handler became `pub(super) fn` so the dispatch in
    reducer.rs can reach it via `domain::handle_xxx(...)`.
  - Tests stay in reducer.rs unchanged — they all dispatch through
    the public `update()` fn, never call handlers directly. (Per
    plan §risk-2.)
  - Each submodule got the minimal `use super::Ctx` /
    `use crate::state::*` imports it needed; cargo check + cargo
    test --bin pass with no new warnings beyond the pre-existing 25.

Generated mechanically by `C:/tmp/split-reducer.js` (Node script
that walks `fn handle_X` blocks, groups by domain, emits per-domain
files with `pub(super)` visibility, then rewrites the dispatch
match in reducer.rs to prefix calls with the domain module).
Script not committed — single-use tool.

Future follow-ups (out of scope per plan §"Out of scope"):
  - Splitting tests per-domain
  - Cross-reducer helper extraction (none anticipated)

866 srv tests pass.
