+++ agentmux-srv/src/reducer.rs
@@ -23,11 +23,18 @@
 // stays pure). The reducer's arm exists only for match
 // exhaustiveness; same pattern as the launcher reducer.
 
-use agentmux_common::ipc::{ClientKind, Command, ErrorCode, Event, LifecyclePhase};
 
-use crate::state::{
-    BlockRecord, ProcessRecord, ProcessState, State, TabRecord, WindowRecord, WorkspaceRecord,
-};
+mod block;
+mod layout;
+mod lifecycle;
+mod snapshot;
+mod tab;
+mod window;
+mod workspace;
+
+use agentmux_common::ipc::{ClientKind, Command, ErrorCode, Event, LifecyclePhase};
+use crate::state::ProcessState;
+use crate::state::State;
 
 /// Per-dispatch context. Currently just an RFC3339 timestamp + the
 /// originating connection's `conn_id` for log correlation.
@@ -40,35 +47,35 @@ pub struct Ctx {
 
 pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
     match cmd {
-        Command::Register { kind, pid, version } => handle_register(state, ctx, kind, pid, version),
-        Command::Goodbye => handle_goodbye(state, ctx),
+        Command::Register { kind, pid, version } => lifecycle::handle_register(state, ctx, kind, pid, version),
+        Command::Goodbye => lifecycle::handle_goodbye(state, ctx),
         Command::Ping { nonce } => {
             let v = state.bump_version();
             vec![Event::Pong { nonce, version: v }]
         }
-        Command::GetSrvSnapshot => handle_get_srv_snapshot(state),
+        Command::GetSrvSnapshot => snapshot::handle_get_srv_snapshot(state),
         Command::GetEvents { .. } => Vec::new(), // intercepted by server; unreachable
-        Command::CreateWorkspace { name } => handle_create_workspace(state, name),
+        Command::CreateWorkspace { name } => workspace::handle_create_workspace(state, name),
         Command::DeleteWorkspace { workspace_id, force } => {
-            handle_delete_workspace(state, workspace_id, force)
+            workspace::handle_delete_workspace(state, workspace_id, force)
         }
-        Command::CreateTab { workspace_id, name } => handle_create_tab(state, workspace_id, name),
-        Command::DeleteTab { workspace_id, tab_id, force } => handle_delete_tab(state, workspace_id, tab_id, force),
+        Command::CreateTab { workspace_id, name } => tab::handle_create_tab(state, workspace_id, name),
+        Command::DeleteTab { workspace_id, tab_id, force } => tab::handle_delete_tab(state, workspace_id, tab_id, force),
         Command::SetActiveTab { workspace_id, tab_id } => {
-            handle_set_active_tab(state, workspace_id, tab_id)
+            tab::handle_set_active_tab(state, workspace_id, tab_id)
         }
         Command::ReorderTab {
             workspace_id,
             tab_id,
             new_index,
-        } => handle_reorder_tab(state, workspace_id, tab_id, new_index),
-        Command::CreateBlock { tab_id, meta } => handle_create_block(state, tab_id, meta),
-        Command::DeleteBlock { tab_id, block_id } => handle_delete_block(state, tab_id, block_id),
+        } => tab::handle_reorder_tab(state, workspace_id, tab_id, new_index),
+        Command::CreateBlock { tab_id, meta } => block::handle_create_block(state, tab_id, meta),
+        Command::DeleteBlock { tab_id, block_id } => block::handle_delete_block(state, tab_id, block_id),
         Command::SetFocusedNode { tab_id, node_id } => {
-            handle_set_focused_node(state, tab_id, node_id)
+            layout::handle_set_focused_node(state, tab_id, node_id)
         }
         Command::SetMagnifiedNode { tab_id, node_id } => {
-            handle_set_magnified_node(state, tab_id, node_id)
+            layout::handle_set_magnified_node(state, tab_id, node_id)
         }
         // Phase E.4.B Phase 5 — layout tree mutation arms. Currently
         // dormant scaffolding (no production callers; Phase 7 migrates
@@ -79,12 +86,12 @@ pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
         Command::LayoutClear {
             tab_id,
             correlation_id,
-        } => handle_layout_clear(state, tab_id, correlation_id),
+        } => layout::handle_layout_clear(state, tab_id, correlation_id),
         Command::LayoutSetTree {
             tab_id,
             new_tree,
             correlation_id,
-        } => handle_layout_set_tree(state, tab_id, new_tree, correlation_id),
+        } => layout::handle_layout_set_tree(state, tab_id, new_tree, correlation_id),
         Command::LayoutInsertNode {
             tab_id,
             node,
@@ -93,7 +100,7 @@ pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
             focus_after,
             magnify_after,
             correlation_id,
-        } => handle_layout_insert_node(
+        } => layout::handle_layout_insert_node(
             state,
             tab_id,
             node,
@@ -107,50 +114,50 @@ pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
             tab_id,
             node_id,
             correlation_id,
-        } => handle_layout_delete_node(state, tab_id, node_id, correlation_id),
+        } => layout::handle_layout_delete_node(state, tab_id, node_id, correlation_id),
         Command::CreateWindow {
             window_id,
             workspace_id,
-        } => handle_create_window(state, window_id, workspace_id),
+        } => window::handle_create_window(state, window_id, workspace_id),
         Command::CloseWindowInternal { window_id } => {
-            handle_close_window_internal(state, window_id)
+            window::handle_close_window_internal(state, window_id)
         }
         Command::SwitchWorkspace {
             window_id,
             workspace_id,
-        } => handle_switch_workspace(state, window_id, workspace_id),
+        } => window::handle_switch_workspace(state, window_id, workspace_id),
         Command::ReorderTabsBulk {
             workspace_id,
             tab_ids,
-        } => handle_reorder_tabs_bulk(state, workspace_id, tab_ids),
+        } => tab::handle_reorder_tabs_bulk(state, workspace_id, tab_ids),
         Command::RenameWorkspace { workspace_id, name } => {
-            handle_rename_workspace(state, workspace_id, name)
+            workspace::handle_rename_workspace(state, workspace_id, name)
         }
-        Command::RenameTab { tab_id, name } => handle_rename_tab(state, tab_id, name),
+        Command::RenameTab { tab_id, name } => tab::handle_rename_tab(state, tab_id, name),
         Command::UpdateWorkspaceMeta {
             workspace_id,
             meta_patch,
-        } => handle_update_workspace_meta(state, workspace_id, meta_patch),
+        } => workspace::handle_update_workspace_meta(state, workspace_id, meta_patch),
         Command::UpdateTabMeta {
             tab_id,
             meta_patch,
-        } => handle_update_tab_meta(state, tab_id, meta_patch),
+        } => tab::handle_update_tab_meta(state, tab_id, meta_patch),
         Command::UpdateBlockMeta {
             block_id,
             meta_patch,
-        } => handle_update_block_meta(state, block_id, meta_patch),
+        } => block::handle_update_block_meta(state, block_id, meta_patch),
         Command::MoveTab {
             tab_id,
             src_workspace_id,
             dst_workspace_id,
             dst_index,
-        } => handle_move_tab(state, tab_id, src_workspace_id, dst_workspace_id, dst_index),
+        } => tab::handle_move_tab(state, tab_id, src_workspace_id, dst_workspace_id, dst_index),
         Command::MoveBlock {
             block_id,
             src_tab_id,
             dst_tab_id,
             dst_index,
-        } => handle_move_block(state, block_id, src_tab_id, dst_tab_id, dst_index),
+        } => block::handle_move_block(state, block_id, src_tab_id, dst_tab_id, dst_index),
         // Anything else is a non-fatal protocol error. Future
         // phases (E.2b tabs, E.3 blocks, E.4 layouts) extend this
         // match by adding new arms above.
@@ -164,1341 +171,6 @@ pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
             }]
         }
     }
-}
-
-fn handle_register(
-    state: &mut State,
-    ctx: &Ctx,
-    kind: ClientKind,
-    pid: u32,
-    version: String,
-) -> Vec<Event> {
-    // Idempotent on duplicate Register from the same PID — preserve
-    // the original record (per launcher's pattern). Accept fresh
-    // Registers only when the PID has no record OR the existing
-    // record is Exited (PID recycled).
-    let prior_state = state.processes.get(&pid).map(|r| r.state);
-    let allow_register = match prior_state {
-        None => true,
-        Some(ProcessState::Exited { .. }) => true,
-        _ => false,
-    };
-    if !allow_register {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::AlreadyRegistered,
-            message: format!("pid {} is already registered with srv", pid),
-            fatal: false,
-            version: v,
-        }];
-    }
-
-    let mut out = Vec::with_capacity(3);
-
-    // Insert the process record.
-    state.processes.insert(
-        pid,
-        ProcessRecord {
-            pid,
-            kind,
-            state: ProcessState::Running,
-            spawned_at: ctx.now_rfc3339.clone(),
-            version: version.clone(),
-        },
-    );
-    let v = state.bump_version();
-    out.push(Event::ProcessSpawned {
-        pid,
-        kind,
-        client_version: version,
-        version: v,
-    });
-
-    // First Register transitions srv to Running. Same lifecycle
-    // pattern as launcher.
-    if state.lifecycle == LifecyclePhase::Starting {
-        state.lifecycle = LifecyclePhase::Running;
-        let v = state.bump_version();
-        out.push(Event::LifecyclePhaseChanged {
-            from: LifecyclePhase::Starting,
-            to: LifecyclePhase::Running,
-            version: v,
-        });
-    }
-
-    let client_id = state.alloc_client_id();
-    let v = state.bump_version();
-    // Sentinel launcher_pid / launcher_version on Registered — IPC
-    // server patches these to the real srv identity before broadcast.
-    out.push(Event::Registered {
-        client_id,
-        launcher_pid: 0,
-        launcher_version: String::new(),
-        version: v,
-    });
-    out
-}
-
-fn handle_goodbye(state: &mut State, ctx: &Ctx) -> Vec<Event> {
-    let Some(pid) = ctx.registered_pid else {
-        return Vec::new();
-    };
-    let Some(record) = state.processes.get_mut(&pid) else {
-        return Vec::new();
-    };
-    if matches!(record.state, ProcessState::Exited { .. }) {
-        return Vec::new();
-    }
-    record.state = ProcessState::Exited { code: 0 };
-    let v = state.bump_version();
-    vec![Event::ProcessExited {
-        pid,
-        code: 0,
-        version: v,
-    }]
-}
-
-fn handle_get_srv_snapshot(state: &mut State) -> Vec<Event> {
-    let v = state.bump_version();
-    let mut workspaces: Vec<(String, String)> = state
-        .workspaces
-        .values()
-        .map(|w| (w.workspace_id.clone(), w.name.clone()))
-        .collect();
-    // Stable ordering for diffability — reducer state is HashMap so
-    // iteration order is non-deterministic.
-    workspaces.sort_by(|a, b| a.0.cmp(&b.0));
-    let mut tabs: Vec<(String, String, String)> = state
-        .tabs
-        .values()
-        .map(|t| (t.tab_id.clone(), t.workspace_id.clone(), t.name.clone()))
-        .collect();
-    tabs.sort_by(|a, b| a.0.cmp(&b.0));
-    let mut active_tabs: Vec<(String, String)> = state
-        .workspaces
-        .values()
-        .filter_map(|w| {
-            w.active_tab_id
-                .as_ref()
-                .map(|t| (w.workspace_id.clone(), t.clone()))
-        })
-        .collect();
-    active_tabs.sort_by(|a, b| a.0.cmp(&b.0));
-    let mut blocks: Vec<(String, String)> = state
-        .blocks
-        .values()
-        .map(|b| (b.block_id.clone(), b.tab_id.clone()))
-        .collect();
-    blocks.sort_by(|a, b| a.0.cmp(&b.0));
-    vec![Event::SrvSnapshot {
-        version: v,
-        lifecycle: state.lifecycle,
-        workspaces,
-        tabs,
-        active_tabs,
-        blocks,
-    }]
-}
-
-/// Phase E.2 — create a new workspace. Reducer assigns the OID
-/// (UUID), inserts into canonical state, emits WorkspaceCreated.
-/// NOT idempotent on retry: each invocation generates a fresh UUID
-/// and inserts a new row, so a saga that double-fires CreateWorkspace
-/// would create two distinct workspaces. Saga-side dedup (correlation
-/// IDs / saga state machine) is responsible for at-most-once delivery
-/// when sagas land in E.5+.
-fn handle_create_workspace(state: &mut State, name: String) -> Vec<Event> {
-    let workspace_id = uuid::Uuid::new_v4().to_string();
-    state.workspaces.insert(
-        workspace_id.clone(),
-        WorkspaceRecord {
-            workspace_id: workspace_id.clone(),
-            name: name.clone(),
-            tab_ids: Vec::new(),
-            active_tab_id: None,
-        },
-    );
-    let v = state.bump_version();
-    vec![Event::WorkspaceCreated {
-        workspace_id,
-        name,
-        version: v,
-    }]
-}
-
-/// Phase E.2 — delete a workspace from canonical state. Idempotent:
-/// deleting a missing workspace is a silent no-op. Cascades to the
-/// workspace's tabs (E.2b) and through to each tab's blocks (E.3):
-/// every tab whose `workspace_id` matches is removed from
-/// `state.tabs`, and each block in those tabs is removed from
-/// `state.blocks`, before the workspace itself goes away. Cascade
-/// events are NOT emitted individually — subscribers observing
-/// `WorkspaceDeleted` are expected to drop dependent state (mirrors
-/// how `wcore::delete_workspace` cascades in SQLite).
-///
-/// The `force` parameter (Step 5 PR 2) is provenance-only: it carries
-/// through to the durable saga log when the saga drives this dispatch
-/// (`force = true`), and is ignored by the reducer's cascade logic.
-/// The reducer is a pure mutator — it must always cascade to keep
-/// in-memory state consistent regardless of whether a saga or a
-/// legacy/internal path is calling.
-fn handle_delete_workspace(
-    state: &mut State,
-    workspace_id: String,
-    _force: bool,
-) -> Vec<Event> {
-    let Some(removed) = state.workspaces.remove(&workspace_id) else {
-        return Vec::new();
-    };
-    for tab_id in &removed.tab_ids {
-        if let Some(tab) = state.tabs.remove(tab_id) {
-            for block_id in &tab.block_ids {
-                state.blocks.remove(block_id);
-            }
-        }
-    }
-    // Phase E.5 — drop window mappings that point at the deleted
-    // workspace AND emit `SrvWindowClosed` for each so the persist
-    // subscriber prunes the SQLite Window row + Client.windowids.
-    // The original cascade (E.5.1+2) was silent, leaving downstream
-    // projections out of sync. (codex P1 follow-up to #619.)
-    let dropped_window_ids: Vec<String> = state
-        .windows
-        .iter()
-        .filter(|(_, w)| w.workspace_id == workspace_id)
-        .map(|(id, _)| id.clone())
-        .collect();
-    for id in &dropped_window_ids {
-        state.windows.remove(id);
-    }
-    let mut events = Vec::with_capacity(1 + dropped_window_ids.len());
-    let v = state.bump_version();
-    events.push(Event::WorkspaceDeleted {
-        workspace_id,
-        version: v,
-    });
-    for window_id in dropped_window_ids {
-        let v = state.bump_version();
-        events.push(Event::SrvWindowClosed { window_id, version: v });
-    }
-    events
-}
-
-/// Phase E.2b — create a tab inside a workspace. Validates the
-/// parent exists; otherwise emits `Event::Error` (non-fatal). On
-/// success: assigns a UUID, appends to the workspace's `tab_ids`,
-/// inserts into `state.tabs`, emits `Event::TabCreated`. If the
-/// workspace had no active tab, the new tab also becomes active
-/// and an `Event::ActiveTabChanged` is emitted alongside.
-///
-/// NOT idempotent on retry (same UUID-assignment caveat as
-/// `handle_create_workspace`).
-fn handle_create_tab(state: &mut State, workspace_id: String, name: String) -> Vec<Event> {
-    let Some(workspace_record) = state.workspaces.get(&workspace_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("CreateTab: workspace not found: {}", workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    // codex P2 #622: auto-generate `tabN` when name is empty,
-    // matching `wcore::create_tab`'s default-naming behaviour. The
-    // counter uses the reducer's tab_ids length + 1 (matching the
-    // old SQLite-side count: tabids.len() + pinnedtabids.len() + 1
-    // — pinnedtabids stays at zero in production since pinning
-    // was removed in E.2c.3b, so reducer-only counting matches).
-    let resolved_name = if name.is_empty() {
-        format!("tab{}", workspace_record.tab_ids.len() + 1)
-    } else {
-        name
-    };
-    let tab_id = uuid::Uuid::new_v4().to_string();
-    state.tabs.insert(
-        tab_id.clone(),
-        TabRecord {
-            tab_id: tab_id.clone(),
-            workspace_id: workspace_id.clone(),
-            name: resolved_name.clone(),
-            block_ids: Vec::new(),
-            focused_node_id: String::new(),
-            magnified_node_id: String::new(),
-            rootnode: None,
-        },
-    );
-    let workspace = state.workspaces.get_mut(&workspace_id).expect("checked");
-    workspace.tab_ids.push(tab_id.clone());
-    let activated = if workspace.active_tab_id.is_none() {
-        workspace.active_tab_id = Some(tab_id.clone());
-        true
-    } else {
-        false
-    };
-    let mut events = Vec::with_capacity(2);
-    let v = state.bump_version();
-    events.push(Event::TabCreated {
-        workspace_id: workspace_id.clone(),
-        tab_id: tab_id.clone(),
-        name: resolved_name,
-        version: v,
-    });
-    if activated {
-        let v2 = state.bump_version();
-        events.push(Event::ActiveTabChanged {
-            workspace_id,
-            tab_id: Some(tab_id),
-            version: v2,
-        });
-    }
-    events
-}
-
-/// Phase E.2b — delete a tab from a workspace. Idempotent: deleting
-/// a missing tab is a silent no-op. If the deleted tab was the
-/// active tab, the workspace's active tab becomes the next tab in
-/// `tab_ids` (or the previous one if the deleted was last; or None
-/// if the workspace is now empty), and an `Event::ActiveTabChanged`
-/// is emitted alongside `Event::TabDeleted`.
-fn handle_delete_tab(
-    state: &mut State,
-    workspace_id: String,
-    tab_id: String,
-    force: bool,
-) -> Vec<Event> {
-    let Some(workspace) = state.workspaces.get_mut(&workspace_id) else {
-        return Vec::new();
-    };
-    let Some(pos) = workspace.tab_ids.iter().position(|t| t == &tab_id) else {
-        return Vec::new();
-    };
-    // Last-tab guard with `force` bypass (round 4 of PR #633).
-    // History:
-    //   * Round 1: saga pre-check → reagent flagged TOCTOU race.
-    //   * Round 2: moved guard to reducer (atomic) → broke CreateTab
-    //     compensation (codex P1 round 2) and Cmd+W keyboard flow
-    //     (codex P1 round 1).
-    //   * Round 3: removed guard entirely; saga keeps soft pre-check
-    //     → codex P2 round 4 re-flagged the TOCTOU race.
-    //   * Round 4 (this): atomic guard with `force: bool` bypass.
-    //     User-facing flows (CloseTab RPC → DeleteTab saga) pass
-    //     `force: false`; compensation paths (`CreateTab` rollback,
-    //     `PromoteBlockToTab.ctx.compensate`) pass `force: true`.
-    //     Frontend keyboard handler `simpleCloseStaticTab` already
-    //     gates pre-RPC, so the reducer rejection is a defense-in-
-    //     depth backstop that catches automation/race paths.
-    if !force && workspace.tab_ids.len() <= 1 {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!(
-                "DeleteTab: refusing to delete the last tab in workspace {} (would leave empty workspace; pass force=true for compensation paths)",
-                workspace_id,
-            ),
-            fatal: false,
-            version: v,
-        }];
-    }
-    workspace.tab_ids.remove(pos);
-    let active_changed = if workspace.active_tab_id.as_deref() == Some(tab_id.as_str()) {
-        let new_active = workspace
-            .tab_ids
-            .get(pos)
-            .or_else(|| pos.checked_sub(1).and_then(|i| workspace.tab_ids.get(i)))
-            .cloned();
-        workspace.active_tab_id = new_active.clone();
-        Some(new_active)
-    } else {
-        None
-    };
-    let removed_tab = state.tabs.remove(&tab_id);
-    // Phase E.3 — cascade to blocks. Subscribers observing TabDeleted
-    // are expected to drop dependent block state (no per-block
-    // BlockDeleted events emitted; mirrors workspace→tabs cascade
-    // semantics).
-    if let Some(tab) = &removed_tab {
-        for block_id in &tab.block_ids {
-            state.blocks.remove(block_id);
-        }
-    }
-    let mut events = Vec::with_capacity(2);
-    let v = state.bump_version();
-    events.push(Event::TabDeleted {
-        workspace_id: workspace_id.clone(),
-        tab_id,
-        version: v,
-    });
-    if let Some(new_active) = active_changed {
-        let v2 = state.bump_version();
-        events.push(Event::ActiveTabChanged {
-            workspace_id,
-            tab_id: new_active,
-            version: v2,
-        });
-    }
-    events
-}
-
-/// Phase E.2b — set a workspace's active tab. No-op if already
-/// active. Errors (non-fatal) if the workspace doesn't exist or the
-/// tab isn't in that workspace's tab list.
-fn handle_set_active_tab(
-    state: &mut State,
-    workspace_id: String,
-    tab_id: String,
-) -> Vec<Event> {
-    let Some(workspace) = state.workspaces.get_mut(&workspace_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("SetActiveTab: workspace not found: {}", workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    if !workspace.tab_ids.iter().any(|t| t == &tab_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!(
-                "SetActiveTab: tab {} not in workspace {}",
-                tab_id, workspace_id
-            ),
-            fatal: false,
-            version: v,
-        }];
-    }
-    if workspace.active_tab_id.as_deref() == Some(tab_id.as_str()) {
-        return Vec::new();
-    }
-    workspace.active_tab_id = Some(tab_id.clone());
-    let v = state.bump_version();
-    vec![Event::ActiveTabChanged {
-        workspace_id,
-        tab_id: Some(tab_id),
-        version: v,
-    }]
-}
-
-/// Phase E.2c.3b — reorder a tab within its workspace's
-/// `tab_ids`. `new_index` is clamped to `tab_ids.len()`. No-op
-/// if the tab is already at that position. Errors (non-fatal) if
-/// the workspace doesn't exist or the tab isn't in its tab list.
-fn handle_reorder_tab(
-    state: &mut State,
-    workspace_id: String,
-    tab_id: String,
-    new_index: u32,
-) -> Vec<Event> {
-    let Some(workspace) = state.workspaces.get_mut(&workspace_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("ReorderTab: workspace not found: {}", workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    let Some(current_pos) = workspace.tab_ids.iter().position(|t| t == &tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!(
-                "ReorderTab: tab {} not in workspace {}",
-                tab_id, workspace_id
-            ),
-            fatal: false,
-            version: v,
-        }];
-    };
-    let len = workspace.tab_ids.len();
-    let target = (new_index as usize).min(len.saturating_sub(1));
-    if current_pos == target {
-        return Vec::new();
-    }
-    let id = workspace.tab_ids.remove(current_pos);
-    workspace.tab_ids.insert(target, id);
-    let v = state.bump_version();
-    vec![Event::TabReordered {
-        workspace_id,
-        tab_id,
-        new_index: target as u32,
-        version: v,
-    }]
-}
-
-/// Phase E.3 — create a block inside a tab. Validates parent tab
-/// exists; otherwise emits `Event::Error` (non-fatal). On success:
-/// assigns a UUID, appends to the tab's `block_ids`, inserts into
-/// `state.blocks`, emits `Event::BlockCreated`.
-///
-/// NOT idempotent on retry (UUID assignment per call); saga-side
-/// dedup is responsible for at-most-once delivery in E.5+.
-fn handle_create_block(
-    state: &mut State,
-    tab_id: String,
-    meta: serde_json::Value,
-) -> Vec<Event> {
-    if !state.tabs.contains_key(&tab_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("CreateBlock: tab not found: {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-    let block_id = uuid::Uuid::new_v4().to_string();
-    state.blocks.insert(
-        block_id.clone(),
-        BlockRecord {
-            block_id: block_id.clone(),
-            tab_id: tab_id.clone(),
-        },
-    );
-    let tab = state.tabs.get_mut(&tab_id).expect("checked");
-    tab.block_ids.push(block_id.clone());
-    let v = state.bump_version();
-    vec![Event::BlockCreated {
-        tab_id,
-        block_id,
-        meta,
-        version: v,
-    }]
-}
-
-/// Phase E.3 — delete a block from a tab. Idempotent: deleting a
-/// missing tab or missing block is a silent no-op.
-fn handle_delete_block(state: &mut State, tab_id: String, block_id: String) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        return Vec::new();
-    };
-    let Some(pos) = tab.block_ids.iter().position(|b| b == &block_id) else {
-        return Vec::new();
-    };
-    tab.block_ids.remove(pos);
-    state.blocks.remove(&block_id);
-    let v = state.bump_version();
-    vec![Event::BlockDeleted {
-        tab_id,
-        block_id,
-        version: v,
-    }]
-}
-
-/// Phase E.4 (Option A) — set the focused layout-node id on a tab.
-/// Errors (non-fatal) if the tab is unknown to the reducer; no-op
-/// short-circuit when the value is already current. Empty `node_id`
-/// clears the field. Bumps the version only on real changes so a
-/// burst of identical sets doesn't churn the event stream.
-fn handle_set_focused_node(state: &mut State, tab_id: String, node_id: String) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("SetFocusedNode: unknown tab {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    if tab.focused_node_id == node_id {
-        return Vec::new();
-    }
-    tab.focused_node_id = node_id.clone();
-    let v = state.bump_version();
-    vec![Event::FocusedNodeChanged {
-        tab_id,
-        node_id,
-        version: v,
-    }]
-}
-
-/// Phase E.4 (Option A) — set the magnified layout-node id on a tab.
-/// Same shape as `handle_set_focused_node`. Empty `node_id` is the
-/// toggle-off / clear case.
-fn handle_set_magnified_node(state: &mut State, tab_id: String, node_id: String) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("SetMagnifiedNode: unknown tab {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    if tab.magnified_node_id == node_id {
-        return Vec::new();
-    }
-    tab.magnified_node_id = node_id.clone();
-    let v = state.bump_version();
-    vec![Event::MagnifiedNodeChanged {
-        tab_id,
-        node_id,
-        version: v,
-    }]
-}
-
-// ── Phase E.4.B Phase 5 — layout tree reducer arms ──────────────────
-//
-// 4 of 11 arms shipped in this PR (clear, set_tree, insert_node,
-// delete_node). Pattern uniform across all 11:
-//   1. Look up tab; emit Event::Error on missing.
-//   2. Mutate `tab.rootnode` via the pure helpers in
-//      `crate::backend::layout::*` (shipped in Phase 4 / PRs #691, #692).
-//   3. Emit the matching Event::Layout* variant carrying the
-//      correlation_id and a fresh version.
-//
-// **No production callers yet** — wcore-direct writers continue to be
-// authoritative until Phase 7 migrates them. These arms are the
-// destination, not the source; tests below exercise them in isolation.
-
-fn handle_layout_clear(
-    state: &mut State,
-    tab_id: String,
-    correlation_id: String,
-) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("LayoutClear: unknown tab {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    tab.rootnode = None;
-    tab.focused_node_id = String::new();
-    tab.magnified_node_id = String::new();
-    let v = state.bump_version();
-    vec![Event::LayoutCleared {
-        tab_id,
-        correlation_id,
-        version: v,
-    }]
-}
-
-fn handle_layout_set_tree(
-    state: &mut State,
-    tab_id: String,
-    new_tree: Option<agentmux_common::LayoutNode>,
-    correlation_id: String,
-) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("LayoutSetTree: unknown tab {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    tab.rootnode = new_tree.clone();
-    // when the tree is wiped, focused/
-    // magnified ids would point at non-existent nodes. Match
-    // `handle_layout_clear`'s contract for the empty-tree case.
-    if new_tree.is_none() {
-        tab.focused_node_id = String::new();
-        tab.magnified_node_id = String::new();
-    }
-    let v = state.bump_version();
-    vec![Event::LayoutTreeReplaced {
-        tab_id,
-        new_tree,
-        correlation_id,
-        version: v,
-    }]
-}
-
-fn handle_layout_insert_node(
-    state: &mut State,
-    tab_id: String,
-    node: agentmux_common::LayoutNode,
-    parent_id: Option<String>,
-    index: Option<usize>,
-    focus_after: bool,
-    magnify_after: bool,
-    correlation_id: String,
-) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("LayoutInsertNode: unknown tab {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    // Three insert paths, in priority order:
-    //   1. Empty tree → promote new node to root.
-    //   2. parent_id given → insert under that specific parent at
-    //      `index` (or append if None). If parent_id is given but
-    //      doesn't resolve to a group node, REJECT with
-    //      Event::Error rather than fall back to the heuristic.
-    //      A silent-fallback path is a consistency hole: the
-    //      emitted event would echo the requested parent_id while
-    //      the actual mutation
-    //      went elsewhere, diverging the persist subscriber and
-    //      replay consumers from the reducer.
-    //   3. parent_id None → `findNextInsertLocation` heuristic.
-    let empty_tree = tab.rootnode.is_none();
-    if empty_tree {
-        // Empty-tree promotion must reject any explicit `parent_id`/
-        // `index` for the same divergence reason as the explicit-
-        // parent path below — the event would echo a target the
-        // tree cannot resolve, and a replay consumer or the persist
-        // subscriber would diverge from the reducer. Per
-        // `srv-phase-e4b-formal-spec-2026-05-03.md` §7.1, both
-        // fields must be `None` for empty-tree promote.
-        if parent_id.is_some() || index.is_some() {
-            let v = state.bump_version();
-            return vec![Event::Error {
-                code: ErrorCode::InvalidCommand,
-                message: format!(
-                    "LayoutInsertNode: empty tree cannot honour explicit parent_id={:?} / index={:?} (tab {})",
-                    parent_id, index, tab_id
-                ),
-                fatal: false,
-                version: v,
-            }];
-        }
-        tab.rootnode = Some(node.clone());
-    } else if let Some(pid) = parent_id.as_deref() {
-        let root = tab.rootnode.as_mut().expect("non-empty checked above");
-        match crate::backend::layout::find_node_by_id_mut(root, pid) {
-            Some(parent_node) if parent_node.data.is_none() => {
-                let len = parent_node.children.len();
-                let target = index.map(|i| i.min(len)).unwrap_or(len);
-                parent_node.children.insert(target, node.clone());
-            }
-            Some(_) => {
-                // parent_id resolves to a leaf — can't host children.
-                let v = state.bump_version();
-                return vec![Event::Error {
-                    code: ErrorCode::InvalidCommand,
-                    message: format!(
-                        "LayoutInsertNode: parent_id {:?} is a leaf node, cannot host children (tab {})",
-                        pid, tab_id
-                    ),
-                    fatal: false,
-                    version: v,
-                }];
-            }
-            None => {
-                let v = state.bump_version();
-                return vec![Event::Error {
-                    code: ErrorCode::InvalidCommand,
-                    message: format!(
-                        "LayoutInsertNode: parent_id {:?} not found in tree (tab {})",
-                        pid, tab_id
-                    ),
-                    fatal: false,
-                    version: v,
-                }];
-            }
-        }
-    } else {
-        let root = tab.rootnode.as_mut().expect("non-empty checked above");
-        crate::backend::layout::insert_node(root, node.clone());
-    }
-    // honour focus_after / magnify_after.
-    // The schema documents these as the side effects callers rely on
-    // for "insert + activate" flows; ignoring them desyncs the snapshot
-    // from the event the caller's handler observed.
-    //
-    // a magnified node must also be the
-    // focused one. Without this, a `magnify_after=true,
-    // focus_after=false` insert leaves `focused_node_id` pointing at
-    // the prior pane while `magnified_node_id` points at the new one
-    // — a UI invariant violation (frontend treats magnify-implies-
-    // focus). Treat magnify as implying focus.
-    if focus_after || magnify_after {
-        tab.focused_node_id = node.id.clone();
-    }
-    if magnify_after {
-        tab.magnified_node_id = node.id.clone();
-    }
-    let v = state.bump_version();
-    vec![Event::LayoutNodeInserted {
-        tab_id,
-        node,
-        // pass the command's
-        // parent_id / index through to the event so subscribers see
-        // what the caller asked for, not a hardcoded `None, None`.
-        // The pure helper currently uses the `findNextInsertLocation`
-        // heuristic and ignores these hints — but the event is the
-        // record of what was *requested*; subscribers can correlate
-        // with the resulting tree by inspecting `node` itself.
-        parent_id,
-        index,
-        correlation_id,
-        version: v,
-    }]
-}
-
-fn handle_layout_delete_node(
-    state: &mut State,
-    tab_id: String,
-    node_id: String,
-    correlation_id: String,
-) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("LayoutDeleteNode: unknown tab {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    // Snapshot pre-delete focus/magnify so we can both detect
-    // direct-target hits AND post-walk for indirect orphaning.
-    let pre_focused = tab.focused_node_id.clone();
-    let pre_magnified = tab.magnified_node_id.clone();
-    let Some(root) = tab.rootnode.as_mut() else {
-        // Empty tree — nothing to delete; idempotent no-op (no event).
-        return Vec::new();
-    };
-    // `backend::layout::delete_node` leaves
-    // root deletion to the caller (returns Ok(()) with the root
-    // unmodified). Detect the root case here and clear the tree
-    // wholesale so the reducer state matches the
-    // `LayoutNodeDeleted` event we emit.
-    if root.id == node_id {
-        tab.rootnode = None;
-    } else if let Err(e) = crate::backend::layout::delete_node(root, &node_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("LayoutDeleteNode: {} (tab {})", e, tab_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-
-    // Two ways focus/magnify ids can reference nodes that no longer
-    // exist in the tree post-delete; both must be cleared:
-    //   - `delete_recursive` collapse-sole-child rewrites a
-    //        parent's id to the promoted child's id. If
-    //        focused/magnified was the parent's original id, that
-    //        id is gone from the tree even though the same physical
-    //        node remains.
-    //   - Deleting a container removes all descendants. If
-    //        focused/magnified was a descendant, it's gone too.
-    // Direct-target match (`pre_focused == node_id`) doesn't catch
-    // either case. Reconcile by walking the post-delete tree and
-    // clearing any focus/magnify id that no longer resolves.
-    let id_resolves = |id: &str| -> bool {
-        if id.is_empty() {
-            return true;
-        }
-        match tab.rootnode.as_ref() {
-            None => false,
-            Some(root) => crate::backend::layout::find_node_by_id(root, id).is_some(),
-        }
-    };
-    let was_focused = !pre_focused.is_empty() && !id_resolves(&pre_focused);
-    let was_magnified = !pre_magnified.is_empty() && !id_resolves(&pre_magnified);
-    if was_focused {
-        tab.focused_node_id = String::new();
-    }
-    if was_magnified {
-        tab.magnified_node_id = String::new();
-    }
-
-    let v = state.bump_version();
-    vec![Event::LayoutNodeDeleted {
-        tab_id,
-        node_id,
-        was_focused,
-        was_magnified,
-        correlation_id,
-        version: v,
-    }]
-}
-
-/// Phase E.5 — record a new window→workspace mapping. Validates
-/// the parent workspace exists; otherwise emits `Event::Error`
-/// (non-fatal). Idempotent on duplicate `window_id`: re-issuing
-/// for the same window updates the workspace pointer if it
-/// changed, or no-ops if identical.
-fn handle_create_window(
-    state: &mut State,
-    window_id: String,
-    workspace_id: String,
-) -> Vec<Event> {
-    if !state.workspaces.contains_key(&workspace_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("CreateWindow: workspace not found: {}", workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-    if let Some(existing) = state.windows.get(&window_id) {
-        if existing.workspace_id == workspace_id {
-            return Vec::new();
-        }
-    }
-    state.windows.insert(
-        window_id.clone(),
-        WindowRecord {
-            window_id: window_id.clone(),
-            workspace_id: workspace_id.clone(),
-        },
-    );
-    let v = state.bump_version();
-    vec![Event::SrvWindowOpened {
-        window_id,
-        workspace_id,
-        version: v,
-    }]
-}
-
-/// Phase E.5 — remove a window's workspace mapping. Idempotent
-/// silent no-op on missing.
-fn handle_close_window_internal(state: &mut State, window_id: String) -> Vec<Event> {
-    if state.windows.remove(&window_id).is_none() {
-        return Vec::new();
-    }
-    let v = state.bump_version();
-    vec![Event::SrvWindowClosed {
-        window_id,
-        version: v,
-    }]
-}
-
-/// Phase E.5 — change which workspace a window points at. Errors
-/// (non-fatal) if the window or destination workspace is unknown.
-/// No-op if the window is already pointing at the destination.
-fn handle_switch_workspace(
-    state: &mut State,
-    window_id: String,
-    workspace_id: String,
-) -> Vec<Event> {
-    if !state.workspaces.contains_key(&workspace_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!(
-                "SwitchWorkspace: destination workspace not found: {}",
-                workspace_id
-            ),
-            fatal: false,
-            version: v,
-        }];
-    }
-    let Some(window) = state.windows.get_mut(&window_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("SwitchWorkspace: window not found: {}", window_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    if window.workspace_id == workspace_id {
-        return Vec::new();
-    }
-    window.workspace_id = workspace_id.clone();
-    let v = state.bump_version();
-    vec![Event::SrvWindowWorkspaceChanged {
-        window_id,
-        workspace_id,
-        version: v,
-    }]
-}
-
-/// Phase E.5.3 — replace a workspace's `tab_ids` with the given
-/// list. Validates the workspace exists and the new list is a
-/// permutation of the current set (same elements, possibly
-/// different order). No-op if identical.
-fn handle_reorder_tabs_bulk(
-    state: &mut State,
-    workspace_id: String,
-    tab_ids: Vec<String>,
-) -> Vec<Event> {
-    // codex P1 #620 carryover: relax membership validation until tab
-    // moves are migrated through the reducer. `MoveTabToWorkspace`
-    // and `PromoteBlockToTab` (planned for PR 4) still write through
-    // wcore without dispatching reducer commands, so the reducer's
-    // view of `workspace.tab_ids` can be stale relative to SQLite.
-    // A subsequent `UpdateTabIds` (now routed through this command)
-    // must not refuse the canonical order just because the reducer
-    // hasn't seen the upstream move yet — that would be a
-    // user-visible regression vs. the prior wcore-direct path.
-    //
-    // Treat the caller's `tab_ids` as authoritative. The remaining
-    // checks are basic sanity: the workspace must exist in the
-    // reducer, and `tab_ids` must not contain duplicates (which would
-    // produce a corrupt persisted ordering with no way for the
-    // subscriber to recover). Length / set comparison against the
-    // reducer's stale view is dropped here; PR 4 reinstates strict
-    // validation once tab moves go through the reducer.
-    if !state.workspaces.contains_key(&workspace_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("ReorderTabsBulk: workspace not found: {}", workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-    {
-        let mut seen: std::collections::HashSet<&String> =
-            std::collections::HashSet::with_capacity(tab_ids.len());
-        for id in &tab_ids {
-            if !seen.insert(id) {
-                let v = state.bump_version();
-                return vec![Event::Error {
-                    code: ErrorCode::InvalidCommand,
-                    message: format!(
-                        "ReorderTabsBulk: tab_ids contains duplicate entry: {}",
-                        id
-                    ),
-                    fatal: false,
-                    version: v,
-                }];
-            }
-        }
-    }
-    if state.workspaces.get(&workspace_id).expect("checked").tab_ids == tab_ids {
-        return Vec::new();
-    }
-    state.workspaces.get_mut(&workspace_id).expect("checked").tab_ids = tab_ids.clone();
-    let v = state.bump_version();
-    vec![Event::TabsReorderedBulk {
-        workspace_id,
-        tab_ids,
-        version: v,
-    }]
-}
-
-/// Phase E.5.5 — move a tab from `src_workspace_id` to
-/// `dst_workspace_id`, inserting at `dst_index` (clamped to dst's
-/// length). Updates the tab's `workspace_id`, removes it from src's
-/// `tab_ids`, inserts into dst's `tab_ids`. If the tab was src's
-/// `active_tab_id`, src's active reverts to its first remaining
-/// tab (or `None` when empty).
-///
-/// Errors when:
-/// * source / dest workspace not found,
-/// * tab not found,
-/// * `tab.workspace_id != src_workspace_id` (caller-side bug),
-/// * `src_workspace_id == dst_workspace_id` (use `ReorderTab` for
-///   intra-workspace reorders — same-workspace moves through this
-///   path would create ambiguity around `dst_index` semantics).
-fn handle_move_tab(
-    state: &mut State,
-    tab_id: String,
-    src_workspace_id: String,
-    dst_workspace_id: String,
-    dst_index: u32,
-) -> Vec<Event> {
-    if src_workspace_id == dst_workspace_id {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: "MoveTab: src and dst workspaces are identical; use ReorderTab".into(),
-            fatal: false,
-            version: v,
-        }];
-    }
-    // Strict validation (Phase E.4 strict-mode flip): the
-    // migration-tolerant lazy-import fallback (codex P1 round-2
-    // #621) was removed once the soak window closed with no
-    // `lazy-import` warnings observed in production. All reducer-
-    // routed paths now keep `state.tabs` and
-    // `state.workspaces[*].tab_ids` consistent with SQLite, so we
-    // can reject unknown tabs and workspace_id mismatches outright.
-    if !state.workspaces.contains_key(&src_workspace_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("MoveTab: src workspace not found: {}", src_workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-    if !state.workspaces.contains_key(&dst_workspace_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("MoveTab: dst workspace not found: {}", dst_workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-    let tab_workspace_id: Option<String> =
-        state.tabs.get(&tab_id).map(|t| t.workspace_id.clone());
-    match tab_workspace_id {
-        None => {
-            let v = state.bump_version();
-            return vec![Event::Error {
-                code: ErrorCode::InvalidCommand,
-                message: format!("MoveTab: tab not found in state: {}", tab_id),
-                fatal: false,
-                version: v,
-            }];
-        }
-        Some(actual_ws) if actual_ws != src_workspace_id => {
-            let v = state.bump_version();
-            return vec![Event::Error {
-                code: ErrorCode::InvalidCommand,
-                message: format!(
-                    "MoveTab: workspace_id mismatch — tab {} belongs to {}, not {}",
-                    tab_id, actual_ws, src_workspace_id
-                ),
-                fatal: false,
-                version: v,
-            }];
-        }
-        Some(_) => {}
-    }
-
-    // Remove from src.
-    let new_src_active_tab_id: Option<String> = {
-        let src = state.workspaces.get_mut(&src_workspace_id).expect("checked");
-        src.tab_ids.retain(|id| id != &tab_id);
-        if src.active_tab_id.as_deref() == Some(tab_id.as_str()) {
-            src.active_tab_id = src.tab_ids.first().cloned();
-        }
-        src.active_tab_id.clone()
-    };
-
-    // Insert into dst at clamped index. Set the moved tab as dst's
-    // new active tab — mirrors wcore::move_tab_to_workspace
-    // behaviour and addresses codex P2 #621 (dst.active_tab_id was
-    // previously left untouched, so a saga-driven tear-off could
-    // produce a destination workspace with no active tab selected).
-    let final_dst_index: u32 = {
-        let dst = state.workspaces.get_mut(&dst_workspace_id).expect("checked");
-        let clamped = (dst_index as usize).min(dst.tab_ids.len());
-        dst.tab_ids.insert(clamped, tab_id.clone());
-        dst.active_tab_id = Some(tab_id.clone());
-        clamped as u32
-    };
-
-    // Update the tab's parent.
-    state
-        .tabs
-        .get_mut(&tab_id)
-        .expect("checked")
-        .workspace_id = dst_workspace_id.clone();
-
-    let v = state.bump_version();
-    vec![Event::TabMoved {
-        tab_id: tab_id.clone(),
-        src_workspace_id,
-        dst_workspace_id,
-        dst_index: final_dst_index,
-        new_src_active_tab_id,
-        new_dst_active_tab_id: Some(tab_id),
-        version: v,
-    }]
-}
-
-/// Phase E.5.5 — move a block from `src_tab_id` to `dst_tab_id` at
-/// `dst_index` (clamped). Updates `block.tab_id`. Cross-tab moves
-/// AND intra-tab repositioning both go through this command (the
-/// caller specifies the destination index regardless).
-///
-/// Errors when source / dest tab missing, block missing, or
-/// `block.tab_id != src_tab_id`.
-fn handle_move_block(
-    state: &mut State,
-    block_id: String,
-    src_tab_id: String,
-    dst_tab_id: String,
-    dst_index: u32,
-) -> Vec<Event> {
-    let validation_error: Option<String> = {
-        if !state.tabs.contains_key(&src_tab_id) {
-            Some(format!("MoveBlock: src tab not found: {}", src_tab_id))
-        } else if !state.tabs.contains_key(&dst_tab_id) {
-            Some(format!("MoveBlock: dst tab not found: {}", dst_tab_id))
-        } else {
-            match state.blocks.get(&block_id) {
-                None => Some(format!("MoveBlock: block not found: {}", block_id)),
-                Some(block) if block.tab_id != src_tab_id => Some(format!(
-                    "MoveBlock: block {} belongs to tab {}, not {}",
-                    block_id, block.tab_id, src_tab_id
-                )),
-                _ => None,
-            }
-        }
-    };
-    if let Some(message) = validation_error {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message,
-            fatal: false,
-            version: v,
-        }];
-    }
-
-    // Special-case intra-tab move: remove and re-insert in the same
-    // tab. The clamp is computed AFTER the removal so dst_index
-    // refers to the post-removal list (matches the spec's "position
-    // in dst.tab_ids AFTER insertion" semantics for cross-tab moves).
-    let final_dst_index: u32 = if src_tab_id == dst_tab_id {
-        let tab = state.tabs.get_mut(&src_tab_id).expect("checked");
-        tab.block_ids.retain(|id| id != &block_id);
-        let clamped = (dst_index as usize).min(tab.block_ids.len());
-        tab.block_ids.insert(clamped, block_id.clone());
-        clamped as u32
-    } else {
-        // Remove from src.
-        state
-            .tabs
-            .get_mut(&src_tab_id)
-            .expect("checked")
-            .block_ids
-            .retain(|id| id != &block_id);
-        // Insert into dst.
-        let dst = state.tabs.get_mut(&dst_tab_id).expect("checked");
-        let clamped = (dst_index as usize).min(dst.block_ids.len());
-        dst.block_ids.insert(clamped, block_id.clone());
-        // Update parent.
-        state.blocks.get_mut(&block_id).expect("checked").tab_id = dst_tab_id.clone();
-        clamped as u32
-    };
-
-    let v = state.bump_version();
-    vec![Event::BlockMoved {
-        block_id,
-        src_tab_id,
-        dst_tab_id,
-        dst_index: final_dst_index,
-        version: v,
-    }]
-}
-
-/// Phase E.5.3 — rename a workspace. Errors if missing; no-op if
-/// the name is unchanged.
-fn handle_rename_workspace(state: &mut State, workspace_id: String, name: String) -> Vec<Event> {
-    let Some(workspace) = state.workspaces.get_mut(&workspace_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("RenameWorkspace: workspace not found: {}", workspace_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    if workspace.name == name {
-        return Vec::new();
-    }
-    workspace.name = name.clone();
-    let v = state.bump_version();
-    vec![Event::WorkspaceRenamed {
-        workspace_id,
-        name,
-        version: v,
-    }]
-}
-
-/// Phase E.5.3 — rename a tab. Errors if missing; no-op if the
-/// name is unchanged.
-fn handle_rename_tab(state: &mut State, tab_id: String, name: String) -> Vec<Event> {
-    let Some(tab) = state.tabs.get_mut(&tab_id) else {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("RenameTab: tab not found: {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    };
-    if tab.name == name {
-        return Vec::new();
-    }
-    tab.name = name.clone();
-    let v = state.bump_version();
-    vec![Event::TabRenamed {
-        tab_id,
-        name,
-        version: v,
-    }]
-}
-
-/// Phase E.5.3 — pass-through validation + emit for workspace
-/// meta updates. The reducer does NOT mutate meta in state (it
-/// doesn't track meta in WorkspaceRecord); the persist subscriber
-/// applies the patch directly to wstore. This keeps the reducer's
-/// state shape unchanged while still routing every meta mutation
-/// through the broadcast bus for observers.
-fn handle_update_workspace_meta(
-    state: &mut State,
-    workspace_id: String,
-    meta_patch: serde_json::Value,
-) -> Vec<Event> {
-    if !state.workspaces.contains_key(&workspace_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!(
-                "UpdateWorkspaceMeta: workspace not found: {}",
-                workspace_id
-            ),
-            fatal: false,
-            version: v,
-        }];
-    }
-    let v = state.bump_version();
-    vec![Event::WorkspaceMetaUpdated {
-        workspace_id,
-        meta_patch,
-        version: v,
-    }]
-}
-
-/// Phase E.5.3 — pass-through for tab meta updates. Same shape as
-/// `handle_update_workspace_meta`.
-fn handle_update_tab_meta(
-    state: &mut State,
-    tab_id: String,
-    meta_patch: serde_json::Value,
-) -> Vec<Event> {
-    if !state.tabs.contains_key(&tab_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("UpdateTabMeta: tab not found: {}", tab_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-    let v = state.bump_version();
-    vec![Event::TabMetaUpdated {
-        tab_id,
-        meta_patch,
-        version: v,
-    }]
-}
-
-/// Phase E.5.3 — pass-through for block meta updates. Same shape.
-fn handle_update_block_meta(
-    state: &mut State,
-    block_id: String,
-    meta_patch: serde_json::Value,
-) -> Vec<Event> {
-    if !state.blocks.contains_key(&block_id) {
-        let v = state.bump_version();
-        return vec![Event::Error {
-            code: ErrorCode::InvalidCommand,
-            message: format!("UpdateBlockMeta: block not found: {}", block_id),
-            fatal: false,
-            version: v,
-        }];
-    }
-    let v = state.bump_version();
-    vec![Event::BlockMetaUpdated {
-        block_id,
-        meta_patch,
-        version: v,
-    }]
 }
 
 #[cfg(test)]
@@ -4383,3 +3055,4 @@ mod tests {
         assert_eq!(state.tabs[&tab_id].focused_node_id, "");
     }
 }
+
