+++ agentmux-srv/src/reducer/lifecycle.rs
@@ -1,0 +1,103 @@
+// Copyright 2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+use agentmux_common::ipc::{Command, ErrorCode, Event};
+
+use crate::state::State;
+
+use super::Ctx;
+
+use agentmux_common::ipc::{ClientKind, LifecyclePhase};
+use crate::state::{ProcessRecord, ProcessState};
+
+pub(super) fn handle_register(
+    state: &mut State,
+    ctx: &Ctx,
+    kind: ClientKind,
+    pid: u32,
+    version: String,
+) -> Vec<Event> {
+    // Idempotent on duplicate Register from the same PID — preserve
+    // the original record (per launcher's pattern). Accept fresh
+    // Registers only when the PID has no record OR the existing
+    // record is Exited (PID recycled).
+    let prior_state = state.processes.get(&pid).map(|r| r.state);
+    let allow_register = match prior_state {
+        None => true,
+        Some(ProcessState::Exited { .. }) => true,
+        _ => false,
+    };
+    if !allow_register {
+        let v = state.bump_version();
+        return vec![Event::Error {
+            code: ErrorCode::AlreadyRegistered,
+            message: format!("pid {} is already registered with srv", pid),
+            fatal: false,
+            version: v,
+        }];
+    }
+
+    let mut out = Vec::with_capacity(3);
+
+    // Insert the process record.
+    state.processes.insert(
+        pid,
+        ProcessRecord {
+            pid,
+            kind,
+            state: ProcessState::Running,
+            spawned_at: ctx.now_rfc3339.clone(),
+            version: version.clone(),
+        },
+    );
+    let v = state.bump_version();
+    out.push(Event::ProcessSpawned {
+        pid,
+        kind,
+        client_version: version,
+        version: v,
+    });
+
+    // First Register transitions srv to Running. Same lifecycle
+    // pattern as launcher.
+    if state.lifecycle == LifecyclePhase::Starting {
+        state.lifecycle = LifecyclePhase::Running;
+        let v = state.bump_version();
+        out.push(Event::LifecyclePhaseChanged {
+            from: LifecyclePhase::Starting,
+            to: LifecyclePhase::Running,
+            version: v,
+        });
+    }
+
+    let client_id = state.alloc_client_id();
+    let v = state.bump_version();
+    // Sentinel launcher_pid / launcher_version on Registered — IPC
+    // server patches these to the real srv identity before broadcast.
+    out.push(Event::Registered {
+        client_id,
+        launcher_pid: 0,
+        launcher_version: String::new(),
+        version: v,
+    });
+    out
+}
+
+pub(super) fn handle_goodbye(state: &mut State, ctx: &Ctx) -> Vec<Event> {
+    let Some(pid) = ctx.registered_pid else {
+        return Vec::new();
+    };
+    let Some(record) = state.processes.get_mut(&pid) else {
+        return Vec::new();
+    };
+    if matches!(record.state, ProcessState::Exited { .. }) {
+        return Vec::new();
+    }
+    record.state = ProcessState::Exited { code: 0 };
+    let v = state.bump_version();
+    vec![Event::ProcessExited {
+        pid,
+        code: 0,
+        version: v,
+    }]
+}
