+++ agentmux-srv/src/reducer/window.rs
@@ -1,0 +1,103 @@
+// Copyright 2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+use agentmux_common::ipc::{Command, ErrorCode, Event};
+
+use crate::state::State;
+
+use super::Ctx;
+
+use crate::state::WindowRecord;
+
+/// Phase E.5 — record a new window→workspace mapping. Validates
+/// the parent workspace exists; otherwise emits `Event::Error`
+/// (non-fatal). Idempotent on duplicate `window_id`: re-issuing
+/// for the same window updates the workspace pointer if it
+/// changed, or no-ops if identical.
+pub(super) fn handle_create_window(
+    state: &mut State,
+    window_id: String,
+    workspace_id: String,
+) -> Vec<Event> {
+    if !state.workspaces.contains_key(&workspace_id) {
+        let v = state.bump_version();
+        return vec![Event::Error {
+            code: ErrorCode::InvalidCommand,
+            message: format!("CreateWindow: workspace not found: {}", workspace_id),
+            fatal: false,
+            version: v,
+        }];
+    }
+    if let Some(existing) = state.windows.get(&window_id) {
+        if existing.workspace_id == workspace_id {
+            return Vec::new();
+        }
+    }
+    state.windows.insert(
+        window_id.clone(),
+        WindowRecord {
+            window_id: window_id.clone(),
+            workspace_id: workspace_id.clone(),
+        },
+    );
+    let v = state.bump_version();
+    vec![Event::SrvWindowOpened {
+        window_id,
+        workspace_id,
+        version: v,
+    }]
+}
+
+/// Phase E.5 — remove a window's workspace mapping. Idempotent
+/// silent no-op on missing.
+pub(super) fn handle_close_window_internal(state: &mut State, window_id: String) -> Vec<Event> {
+    if state.windows.remove(&window_id).is_none() {
+        return Vec::new();
+    }
+    let v = state.bump_version();
+    vec![Event::SrvWindowClosed {
+        window_id,
+        version: v,
+    }]
+}
+
+/// Phase E.5 — change which workspace a window points at. Errors
+/// (non-fatal) if the window or destination workspace is unknown.
+/// No-op if the window is already pointing at the destination.
+pub(super) fn handle_switch_workspace(
+    state: &mut State,
+    window_id: String,
+    workspace_id: String,
+) -> Vec<Event> {
+    if !state.workspaces.contains_key(&workspace_id) {
+        let v = state.bump_version();
+        return vec![Event::Error {
+            code: ErrorCode::InvalidCommand,
+            message: format!(
+                "SwitchWorkspace: destination workspace not found: {}",
+                workspace_id
+            ),
+            fatal: false,
+            version: v,
+        }];
+    }
+    let Some(window) = state.windows.get_mut(&window_id) else {
+        let v = state.bump_version();
+        return vec![Event::Error {
+            code: ErrorCode::InvalidCommand,
+            message: format!("SwitchWorkspace: window not found: {}", window_id),
+            fatal: false,
+            version: v,
+        }];
+    };
+    if window.workspace_id == workspace_id {
+        return Vec::new();
+    }
+    window.workspace_id = workspace_id.clone();
+    let v = state.bump_version();
+    vec![Event::SrvWindowWorkspaceChanged {
+        window_id,
+        workspace_id,
+        version: v,
+    }]
+}
